# slightly faster l1 attention cuda kernels