I have made the following changes:

1. Included shared memory usage to cache query and key vectors and hence reduced global memory accesses. This is done in the forward pass of `l1attn_cuda_forward_kernelX`
```
__shared__ scalar_t q_shared[32];
__shared__ scalar_t k_shared[32];
```

Also modified the mail computation loop to utilise the shared memory:

```
if(o < width) {
    q_shared[tix] = q[b][t][h][o];
    k_shared[tix] = k[b][s][h][o];
    f += abs(q_shared[tix] - k_shared[tix]); 
}
```

2. In the forwad pass of the `l1attn_cuda_forward_kernel16` kernel,

the memory loading pattern is modified:

```
for (int i = 0; i < 2; i++) {
    qc[cu + i*8][cw] = q[b][h][t + i*8][cw];
    kc[cu + i*8][cw] = k[b][h][s + i*8][cw];
}
```

3. In the backward pass of the ` l1attn_cuda_backward_kernel` I added shared memory for q and k:

```
__shared__ scalar_t q_shared[32];
__shared__ scalar_t k_shared[32];
```

and then, modified the computation to use shared memory:

```
q_shared[tix] = q[b][w][h][r];
scalar_t qq = q_shared[tix];

```


4. In the backward pass of the ` l1attn_cuda_backward_kernel16` I modified memory loading for better coalescing:

```
for (int i = 0; i < 2; i++) {
    qc[cr + i*8][cw] = q[b][h][t + i*8][cw];
    kc[cr + i*8][cw] = k[b][h][s + i*8][cw];
}

```
---

After making these changes, I wanted to test the improvements in speed, so used two identical cloud VMs running the following identical specs:

OS: Ubuntu 20.04.6 LTS
GPU: NVIDIA Tesla T4
CPU: Intel Xeon Platinum 8259CL (8) @ 3.100GHz
RAM: 31628MiB
(above details obtained from running neofetch)

The benchmark_l1_attn.py script was then used to test the improvement in optimisations and the following results are obtained:

| Configuration | Original (ms) | Optimised (ms) |
|---------------|---------------|----------------|
|1x128x8x64|0.21|0.20|
|1x512x8x64|3.86|3.83|
|1x1024x8x64|15.47|15.30|
|8x128x8x64|1.95|1.96|
|8x512x8x64|31.22|31.14|
|8x1024x8x64|124.57|123.07|
|32x1024x8x64|504.52|498.41|


I feel this improvement is more noticeable with larger batches and would love to address any issues this code might cause. My email is koushik.sss@outlook.com