@tlh24 2024-08-06
Thanks for looking into this!  

I ran some tests yesterday, and the edits yield performance changes entirely within measurement noise, alas.  Looking at the code, this makes sense, as the introduced variables will be optimized out by the compiler.